+++ packages/utils/src/styles/tailwind.config.js
@@ -34,7 +34,7 @@ export default {
         inherit: 'inherit',
       },
       zIndex: {
-        10000: 10000,
+        max: 2147483647,
       },
       lineHeight: {
         inherit: 'inherit',